Work on ListOptions   Added Vendor, and converted IDs to []int64 to match the corresponding ID changes committed prior.
Vendor added as needed base functionality for product selection.

+++ inventory_item_test.go
@@ -71,7 +71,7 @@ func TestInventoryItemsListWithIDs(t *testing.T) {
 	)
 
 	options := ListOptions{
-		IDs: []int{1, 2},
+		IDs: []int64{1, 2},
 	}
 
 	items, err := client.InventoryItem.List(options)
